+++ tests/testthat/test-inherit.R
@@ -21,4 +21,53 @@ test_that("configurations can inherit from multiple other configurations", {
   expect_identical(config$log, TRUE)
 })
 
+test_that("inheritance can be an expression", {
+  expect_equal(
+    config::get("letter", file = "config-inherits-expr.yml"),
+    "a"
+  )
+  withr::with_envvar(c(CONFIG = "b"), {
+    expect_equal(
+      config::get("letter", config = "b", file = "config-inherits-expr.yml"),
+      "b"
+    )
+  })
+  withr::with_envvar(c(CONFIG = "b"), {
+    expect_equal(
+      config::get("letter", config = "shinyapps", file = "config-inherits-expr.yml"),
+      "b"
+    )
+  })
+  withr::with_envvar(c(CONFIG = "b"), {
+    expect_equal(
+      config::get("letter", config = "noinherit", file = "config-inherits-expr.yml"),
+      "c"
+    )
+  })
 
+  yml_file <- "config-inherits-expr.yml"
+  expect_identical(
+    config::get(file = "config-inherits-expr.yml"),
+    structure(
+      list(
+        letter = "a"
+      ),
+      config = "default",
+      class = "config",
+      file = normalizePath(yml_file)
+    )
+  )
+  withr::with_envvar(c(CONFIG = "b"), {
+    expect_identical(
+    config::get(file = "config-inherits-expr.yml"),
+    structure(
+      list(
+        letter = "a"
+      ),
+      config = "default",
+      class = "config",
+      file = normalizePath(yml_file)
+    )
+  )
+  })
+})